// src/lib/types/appwrite.d.ts
import type { Models } from 'node-appwrite';

// This file is auto-generated by the Appwrite CLI. 
// You can regenerate it by running `appwrite types src/lib/types`.

export enum Roles {
    PARTICIPANT = "participant",
    VOLUNTEER = "volunteer",
    COORDINATOR = "coordinator"
}

export enum FraudStatus {
    CLEAR = "clear",
    SUSPICIOUS = "suspicious",
    CONFIRMED = "confirmed"
}

export enum Priority {
    HIGH = "high",
    MEDIUM = "medium",
    LOW = "low"
}

export enum ReviewStatus {
    APPROVED = "approved",
    DENIED = "denied",
    UNDER_REVIEW = "under-review",
    CANCELED = "canceled"
}

export enum Role {
    STUDENT = "student",
    EDUCATOR = "educator",
    PARENT = "parent",
    OTHER = "other"
}

export enum Classes {
    REMEDIAL = "remedial",
    REGULAR = "regular",
    HONORS = "honors"
}

export enum GradYear {
    2027 = "2027",
    2028 = "2028",
    2029 = "2029",
    2030 = "2030",
    OTHER = "other"
}

export type User = Models.Row & {
    roles: Roles[] | null;
    currency: number;
    fraudStatus: FraudStatus;
    approved: boolean;
    approvalDocumentFile: string | null;
    birthDate: string | null;
    name: string | null;
}

export type GrantApplication = Models.Row & {
}

export type VolunteerTask = Models.Row & {
    taskTitle: string | null;
    description: string | null;
    priority: Priority | null;
    instructions: string | null;
}

export type Signature = Models.Row & {
    users: User;
    signatureFile: string | null;
    date: string | null;
}

export type VolunteerReport = Models.Row & {
    totalHours: number;
    reviewStatus: ReviewStatus;
    reviewerComments: string | null;
    description: string;
    reviewer: User;
    volunteer: User;
    documents: string[] | null;
}

export type ShopItem = Models.Row & {
    public: boolean;
    releaseOn: string | null;
    title: string | null;
}

export type InterestSurveyResponse = Models.Row & {
    email: string;
    role: Role;
    classes: Classes | null;
    gradYear: GradYear | null;
    struggles: string | null;
}